Добавляем загрузчики для css и png + какие-то изменения в компонентах - не обращать внимания, все равно все под замену
