+++ webpack.config.js
@@ -10,6 +10,7 @@ module.exports = {
     contentBase: path.resolve(__dirname, `public`),
     open: false,
     port: 1337,
+    historyApiFallback: true //нужно для работы React-router
   },
   module: {
     rules: [
@@ -18,7 +19,7 @@ module.exports = {
         exclude: /node_modules/,
         use: {
           loader: `babel-loader`,
-        },
+        }
       },
       // настреиваем загрузчики для css и png для работы с пакетом leaflet
       {
@@ -36,44 +37,3 @@ module.exports = {
   },
   devtool: `source-map`,
 };
-
-/* Old file
-const path = require(`path`);
-
-module.exports = {
-  entry: `./src/index.js`,
-  output: {
-    filename: `bundle.js`,
-    path: path.resolve(__dirname, `public`)
-  },
-  devServer: {
-    contentBase: path.resolve(__dirname, `public`),
-    open: false,
-    port: 1337,
-    historyApiFallback: true
-  },
-  module: {
-    rules: [
-      {
-        test: /\.(js|jsx)$/,
-        exclude: /node_modules/,
-        use: {
-          loader: `babel-loader`,
-        },
-      },
-      {
-        test: /\.css$/i,
-        use: [`style-loader`, `css-loader`]
-      },
-      {
-        test: /\.png$/i,
-        use: [`file-loader`]
-      }
-    ],
-  },
-  resolve: {
-    extensions: [`.js`, `.jsx`]
-  },
-  devtool: `source-map`,
-};
-*/
